Add files via upload
The Age Calculator is a simple web application that allows users to determine their age based on their date of birth. By entering a date of birth into an input field and clicking the "Calculate" button, users can instantly see their current age displayed in years. This project is an excellent example of integrating HTML, CSS, and JavaScript to create a functional and interactive web tool